pyproject.toml
[project]
name = "py-chess"
version = "0.1.0"
description = "A HATEOAS-based chess server with real-time updates"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "starlette",

    # Chess logic
    "python-chess>=1.11",

    # Database
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "psycopg2-binary",
    "alembic>=1.14",

    # Templating
    "jinja2>=3.1",

    # Authentication
    "fastapi-sso>=0.16",
    "python-jose[cryptography]",
    "types-python-jose",

    # SSE for real-time updates
    "sse-starlette>=2.1",

    # Utilities
    "python-dotenv",
    "python-multipart",
    "pydantic>=2.0",
    "httpx>=0.28",
    "watchfiles",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.25",
    "ruff>=0.8",
    "mypy>=1.13",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
